jwt spring social login

+++ main-Login-2/src/main/java/main/spring/login/repository/UserRepository.java
@@ -3,7 +3,7 @@ package main.spring.login.repository;
 import main.spring.login.entity.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<UserEntity, Integer> {
+public interface UserRepository extends JpaRepository<UserEntity, Long> {
 
 
     Boolean existsByUsername(String username);
